feat(number): change folder InputNumber to Number
