+++ packages/react/src/Form/Number/Number.mdx
@@ -3,9 +3,9 @@ import { ArgsTable } from "@storybook/addon-docs";
 import * as NumberStories from "./Number.stories";
 import * as NumberInputStories from "./NumberInput.stories";
 
-# Input Number
+# Number Field
 
-The Number component exposes a native HTML number input. Therefore, it accepts a specific range of characters :
+The Number field component exposes a native HTML number input. Therefore, it accepts a specific range of characters :
 
 - number [0-9]
 - "e" for exposant
@@ -100,8 +100,6 @@ export default NumberInputReturn;
 ```
 
 <Canvas of={NumberInputStories.NumberInputStory} />
-
-<ArgsTable story="NumberInput" />
 
 ### NumberInput Required
