+++ packages/react/src/Form/Number/Number.stories.tsx
@@ -1,29 +1,25 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { NumberInput } from "./NumberInput";
-import { MessageTypes } from "../core";
+import { Number } from "./Number";
 
 const MODIFIERS = ["", "required", "disabled"];
 
-const meta: Meta<typeof NumberInput> = {
-  component: NumberInput,
+const meta: Meta<typeof Number> = {
+  component: Number,
   title: "Components/Form/Input/Number",
   argTypes: { onChange: { action: "onChange" } },
 };
 
 export default meta;
 
-type StoryProps = Omit<
-  React.ComponentProps<typeof NumberInput>,
-  "classModifier"
-> & {
+type StoryProps = Omit<React.ComponentProps<typeof Number>, "classModifier"> & {
   classModifier: string[];
 };
 type Story = StoryObj<StoryProps>;
 
-export const NumberInputStory: Story = {
-  name: "NumberInput",
+export const NumberStory: Story = {
+  name: "Number",
   render: ({ classModifier, onChange, ...args }) => (
-    <NumberInput
+    <Number
       classModifier={classModifier.join(" ")}
       onChange={onChange}
       {...args}
@@ -39,19 +35,12 @@ export const NumberInputStory: Story = {
     disabled: false,
     autoFocus: false,
     className: "",
-    label: "Your name",
-    message: "",
-    helpMessage: "Aide à la saisie",
-    forceDisplayMessage: false,
-    messageType: MessageTypes.error,
-    classNameContainerLabel: "col-md-2",
-    classNameContainerInput: "col-md-10",
     "aria-disabled": false,
   },
   argTypes: {
     classModifier: {
       options: MODIFIERS,
-      control: { type: "multi-select" },
+      control: { type: "inline-check" },
     },
   },
 };
