+++ packages/react/src/Form/Number/NumberInput.stories.tsx
@@ -1,25 +1,29 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Number } from "./Number";
+import { NumberInput } from "./NumberInput";
+import { MessageTypes } from "../core";
 
 const MODIFIERS = ["", "required", "disabled"];
 
-const meta: Meta<typeof Number> = {
-  component: Number,
+const meta: Meta<typeof NumberInput> = {
+  component: NumberInput,
   title: "Components/Form/Input/Number",
   argTypes: { onChange: { action: "onChange" } },
 };
 
 export default meta;
 
-type StoryProps = Omit<React.ComponentProps<typeof Number>, "classModifier"> & {
+type StoryProps = Omit<
+  React.ComponentProps<typeof NumberInput>,
+  "classModifier"
+> & {
   classModifier: string[];
 };
 type Story = StoryObj<StoryProps>;
 
-export const NumberStory: Story = {
-  name: "Number",
+export const NumberInputStory: Story = {
+  name: "NumberInput",
   render: ({ classModifier, onChange, ...args }) => (
-    <Number
+    <NumberInput
       classModifier={classModifier.join(" ")}
       onChange={onChange}
       {...args}
@@ -35,12 +39,19 @@ export const NumberStory: Story = {
     disabled: false,
     autoFocus: false,
     className: "",
+    label: "Your name",
+    message: "",
+    helpMessage: "Aide à la saisie",
+    forceDisplayMessage: false,
+    messageType: MessageTypes.error,
+    classNameContainerLabel: "col-md-2",
+    classNameContainerInput: "col-md-10",
     "aria-disabled": false,
   },
   argTypes: {
     classModifier: {
       options: MODIFIERS,
-      control: { type: "multi-select" },
+      control: { type: "inline-check" },
     },
   },
 };
